+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/controller/SimulationController.kt
@@ -1,8 +1,10 @@
-package app.urbanflo.urbanflosumoserver
+package app.urbanflo.urbanflosumoserver.controller
 
-import app.urbanflo.urbanflosumoserver.responses.NewSimulationResponse
-import app.urbanflo.urbanflosumoserver.responses.SimulationInfo
-import app.urbanflo.urbanflosumoserver.responses.SumoNetwork
+import app.urbanflo.urbanflosumoserver.SimulationInstance
+import app.urbanflo.urbanflosumoserver.SimulationStep
+import app.urbanflo.urbanflosumoserver.model.NewSimulationResponse
+import app.urbanflo.urbanflosumoserver.model.SimulationInfo
+import app.urbanflo.urbanflosumoserver.model.SumoNetwork
 import app.urbanflo.urbanflosumoserver.storage.StorageBadRequestException
 import app.urbanflo.urbanflosumoserver.storage.StorageException
 import app.urbanflo.urbanflosumoserver.storage.StorageService
