+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/model/ModelUtils.kt
@@ -1,0 +1,3 @@
+package app.urbanflo.urbanflosumoserver.model
+
+typealias SumoEntityId = String
