--- src/main/kotlin/app/urbanflo/urbanflosumoserver/responses/EntityUtils.kt
@@ -1,3 +1,0 @@
-package app.urbanflo.urbanflosumoserver.responses
-
-typealias SumoEntityId = String
