--- src/main/kotlin/app/urbanflo/urbanflosumoserver/responses/NewSimulationResponse.kt
@@ -1,3 +1,0 @@
-package app.urbanflo.urbanflosumoserver.responses
-
-data class NewSimulationResponse(val id: String)
