--- src/main/kotlin/app/urbanflo/urbanflosumoserver/responses/SimulationInfo.kt
@@ -1,3 +1,0 @@
-package app.urbanflo.urbanflosumoserver.responses
-
-data class SimulationInfo(val id: String)
