+++ src/main/kotlin/app/urbanflo/urbanflosumoserver/storage/StorageService.kt
@@ -2,11 +2,9 @@ package app.urbanflo.urbanflosumoserver.storage
 
 import app.urbanflo.urbanflosumoserver.SimulationId
 import app.urbanflo.urbanflosumoserver.SimulationInstance
-import app.urbanflo.urbanflosumoserver.responses.SimulationInfo
-import app.urbanflo.urbanflosumoserver.responses.SumoNetwork
+import app.urbanflo.urbanflosumoserver.model.SimulationInfo
+import app.urbanflo.urbanflosumoserver.model.SumoNetwork
 import org.springframework.core.io.Resource
-import org.springframework.web.multipart.MultipartFile
-import java.util.*
 
 interface StorageService {
     fun store(network: SumoNetwork): SimulationId
